chore: updated tests with new structure
