+++ src/v0/destinations/google_adwords_remarketing_lists/util.test.js
@@ -1,4 +1,4 @@
-const { populateIdentifiers, responseBuilder } = require('./util');
+const { populateIdentifiers, responseBuilder, getOperationAudienceId } = require('./util');
 const { API_VERSION } = require('./config');
 const accessToken = 'abcd1234';
 const developerToken = 'ijkl9101';
@@ -29,7 +29,7 @@ const body = {
 const baseDestination = {
   Config: {
     rudderAccountId: '258Yea7usSKNpbkIaesL9oJ9iYw',
-    listId: '7090784486',
+    audienceId: '7090784486',
     customerId: '7693729833',
     loginCustomerId: '',
     subAccount: false,
@@ -150,7 +150,7 @@ describe('GARL utils test', () => {
         developerToken,
         body,
         baseDestination,
-        message,
+        getOperationAudienceId(baseDestination.Config.audienceId, message),
         consentBlock,
       );
       expect(response).toEqual(expectedResponse);
@@ -166,7 +166,7 @@ describe('GARL utils test', () => {
           developerToken,
           body,
           destination2,
-          message,
+          getOperationAudienceId(baseDestination.Config.audienceId, message),
           consentBlock,
         );
         expect(response).toEqual();
@@ -178,13 +178,13 @@ describe('GARL utils test', () => {
     it('Should throw error if operationAudienceId is not defined', () => {
       try {
         const destination1 = Object.create(baseDestination);
-        destination1.Config.listId = '';
+        destination1.Config.audienceId = '';
         const response = responseBuilder(
           accessToken,
           developerToken,
           body,
           destination1,
-          message,
+          getOperationAudienceId(baseDestination.Config.audienceId, message),
           consentBlock,
         );
         expect(response).toEqual();
@@ -196,7 +196,13 @@ describe('GARL utils test', () => {
 
   describe('populateIdentifiers function tests', () => {
     it('Should hash and return identifiers for a given list of attributes', () => {
-      const identifier = populateIdentifiers(attributeArray, baseDestination);
+      const { typeOfList, isHashRequired } = baseDestination.Config;
+      const identifier = populateIdentifiers(
+        attributeArray,
+        baseDestination,
+        typeOfList,
+        isHashRequired,
+      );
       expect(identifier).toEqual(hashedArray);
     });
   });
