+++ test/integrations/destinations/google_adwords_remarketing_lists/router/audience.ts
@@ -4,7 +4,7 @@ import { generateGoogleOAuthMetadata } from '../../../testUtils';
 const destination: Destination = {
   Config: {
     rudderAccountId: '258Yea7usSKNpbkIaesL9oJ9iYw',
-    listId: '7090784486',
+    audienceId: '7090784486',
     customerId: '7693729833',
     loginCustomerId: '',
     subAccount: false,
